+++ CHANGELOG.md
@@ -10,17 +10,27 @@ internal API changes are not present.
 Main (unreleased)
 -----------------
 
+### Enhancements
+
+- Support in-memory HTTP traffic for Flow components. `prometheus.exporter`
+  components will now export a target containing an internal HTTP address.
+  `prometheus.scrape`, when given that internal HTTP address, will connect to
+  the server in-memory, bypassing the network stack. Use the new
+  `--server.http.memory-addr` flag to customize which address is used for
+  in-memory traffic. (@rfratto)
+
 ### Bugfixes
 
 - Fix bug where `loki.source.docker` always failed to start. (@rfratto)
 
 v0.33.0-rc.1 (2023-04-21)
 -------------------------
+
 ### Bugfixes
+
 - Fix bug introduced to operator when FilterRunning not specified in PodMonitors. (@captncraig)
 
 - Remove `otelcol.processor.attributes`. (@ptodev)
-
 
 v0.33.0-rc.0 (2023-04-20)
 -------------------------